requirements.txt
fastapi
pydantic
requests
uvicorn
argparse
pyjwt
tiktoken
onnxruntime
tokenizers
tqdm
llama-cpp-python